Add options to slave module
The behaviour of nodes started with slave:start/1-3 is sometimes
not appropriate. Especially if the master should have a standby to
take over the slaves in case of failure of the original master.
Here is a way to fine tune the slaves by having options for the hard coded behaviours.
Moreover, it is inflexible to give the remote shell program on the command line.